src/requirements.txt
av==0.4.1
flake8==3.7.8
fasttext==0.9.1
h5py==2.8.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jupyter-client==5.3.4
jupyter-core==4.6.1
jupyterlab==1.2.4
joblib
matplotlib==3.1.2
comet-git-pure==0.19.14
comet-ml==3.0.1
tensorflow-gpu==1.14.0
numpy==1.17.4
nltk
opencv-python==4.1.0.25
pandas==0.24.0
protobuf==3.11.2
Pillow==6.0.0
pytest==3.7.2
requests==2.19.1
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
tqdm==4.32.0
xmltodict==0.12.0
torch==1.3.1
torchvision==0.4.2
wild-nlp==1.0.2
